+++ src/App.vue
@@ -1,8 +1,6 @@
 <template>
   <div class="app">
-    <SiteHeader class="header"></SiteHeader>
-    <v-spinner v-if="loading"></v-spinner>
-    <router-view v-else></router-view>
+    <app-layout />
     <vue-griddle />
   </div>
 </template>
@@ -12,31 +10,21 @@ import { mapState, mapActions } from 'vuex';
 import i18n from '@/locale/i18n';
 import { auth } from '@/config/firebaseConfig';
 
+import AppLayout from './components/AppLayout.vue';
+
 export default {
   name: 'App',
 
-  metaInfo() {
-    return {
-      title: ` ${i18n.t('general.project')} | ${i18n.t('general.owner')}`,
-    };
-  },
-
   components: {
-    SiteHeader: () => import('@/components/Navigation/SiteHeader.vue'),
+    AppLayout,
   },
 
   computed: {
-    ...mapState(['user', 'loading', 'LS_MODE']),
+    ...mapState(['LS_MODE']),
 
     isDev() {
       return import.meta.env.NODE_ENV !== 'production';
     },
-  },
-
-  created() {
-    if (document.querySelector('#spinner')) {
-      document.querySelector('#spinner').remove();
-    }
   },
 
   mounted() {