+++ src/db/Objective/Objective.js
@@ -1,6 +1,6 @@
 import { db } from '@/config/firebaseConfig';
 import props from './props';
-import { validateCreateProps, createDocument, validateUpdateProps, updateDocument, deleteDocument } from '../common';
+import { validateCreateProps, createDocument, validateUpdateProps, updateDocument } from '../common';
 import KeyResult from '../KeyResult';
 
 const collection = db.collection('objectives');
@@ -29,14 +29,4 @@ const archive = (id) => {
 };
 const restore = (id) => update(id, { archived: false });
 
-const deleteDeep = async (id) => {
-  // Delete affected key results
-  db.collection('keyResults')
-    .where('objective', '==', collection.doc(id))
-    .get()
-    .then(({ docs }) => docs.forEach(({ ref }) => KeyResult.deleteDeep(ref.id)));
-
-  return deleteDocument(update, collection.doc(id));
-};
-
-export default { create, update, archive, restore, deleteDeep };
+export default { create, update, archive, restore };